fix. AddDateFragment & Alarm
Alarm 등록 코드 분리
기기 재부팅 시 알람을 등록해주는 코드 주석처리
